Completely disable redis data if we are not saving
If redis was previously set to save data (to dump.rdb), it would
still load this on startup if subsequently set to not save.

This is broken behaviour, and can cause problems when limiting its
memory. It also will load old data even if it hasn't been saved.

Turning it off completely is more sane, lets do that instead.